Renamed package
to comply with checkstyle package name rules
