--- src/compute_engine_components_api/ComputeRequest.java
@@ -1,5 +1,0 @@
-package compute_engine_components_api;
-
-public class ComputeRequest {
-
-}
